backend/requirements.txt
aiomysql==0.3.2
aiosmtplib==3.0.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.30.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
passlib==1.7.4
psycopg2==2.9.11
pyasn1==0.4.8
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.7.1
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
resend==2.21.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.36
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
